requirements.txt
torch
torchvision
timm
numpy
pillow
scikit-learn
